+++ test/tracker_test.rb
@@ -21,4 +21,19 @@ class TrackerTest < Minitest::Test
     event = Ahoy::Event.last
     assert_equal user.id, event.user_id
   end
+
+  def test_track_user_id
+    ahoy = Ahoy::Tracker.new
+    assert ahoy.track("Some event", {some_prop: true}, user_id: 123)
+
+    event = Ahoy::Event.last
+    assert_equal 123, event.user_id
+    assert_equal 123, event.visit.user_id
+
+    assert ahoy.track("Some event", {some_prop: true}, user_id: 456)
+    event = Ahoy::Event.last
+    assert_equal 456, event.user_id
+    # user id not changed
+    assert_equal 123, event.visit.user_id
+  end
 end